#55 - Refactor authentication routes: consolidate login and registration logic into AuthRouter and remove deprecated LoginRouter
